% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{allMinimaxDesigns}
\alias{allMinimaxDesigns}
\title{Find the minimax design for a two-stage trial}
\usage{
allMinimaxDesigns(p, ntot, pearly = 0.1, alpha = 0.1)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2)}

\item{ntot}{scalar containing the total sample size planned for the trial (n1+n2)}

\item{pearly}{desired probability of early stopping (default = .1).}

\item{alpha}{desired significance level (default = .1).}
}
\description{
Calculates the minimax probability for each combination of n1 and n2 for a given total n
}
\examples{
allMinimaxDesigns(c(.8, .2), 36)
allMinimaxDesigns(c(.7, .3), 40, pearly = .08, alpha=.1)
}
