% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{best_designs}
\alias{best_designs}
\title{Finds the minimax and optimal designs for a two-stage trial}
\usage{
best_designs(p, ntot, pearly = 0.1, alpha = 0.1)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2)}

\item{ntot}{scalar containing the total sample size planned for the trial (n1+n2)}

\item{pearly}{desired probability of early stopping (default = .1).}

\item{alpha}{desired significance level (default = .1).}
}
\description{
Finds the minimax and optimal designs for a two-stage trial
}
\examples{
best_designs(c(.8, .2), 36)
best_designs(c(.7, .3), 40, pearly = .08, alpha=.1)
}
