% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{critical_values}
\alias{critical_values}
\title{Find critical values for decision making during the one-stage or two-stage trial}
\usage{
critical_values(x, alpha = 0.1, p_early = 0.1)
}
\arguments{
\item{alpha}{desired significance level (default is 0.1).}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2).}

\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2) under the null hypothesis.}

\item{pearly}{desired probability of early stopping (default is 0.1).
Not necessary for determining critical values for the one-stage design.}
}
\description{
Finds the critical number of successes necessary to reject 
the null hypothesis in the one- and two-stage designs, 
as well as the number of Stage 1 successes neccesary to 
continue to Stage 2 in the two-stage design.
}
\details{
For the one-stage design:  
Finds the value of s, the minimum number of successes to reject the 
null hypothesis, from the Binomial model with significance level alpha.

For the two-stage design:
Finds r1, the minimum number of Stage 1 successes to continue to 
Stage 2, and r2, the minimum number of Stage 2 successes to 
reject the null hypothesis.  Find r1 so that the probability 
of early stopping is less than or equal to pearly. 
Find r2 from Binomial model with no early stopping
and significance level alpha.
}
\examples{
critical_values(n=36, p=0.2, alpha=.1)
critical_values(n=c( 5, 31), p=c(.8,.2), pearly=.1, alpha=.1)
}
