\name{ctp2nc-package}
\alias{ctp2nc-package}
\alias{ctp2nc}
\docType{package}
\title{
\packageTitle{ctp2nc}
}
\description{
\packageDescription{ctp2nc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ctp2nc}
\packageIndices{ctp2nc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ctp2nc}

Maintainer: \packageMaintainer{ctp2nc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
