% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r, R/snb.r
\name{dsnb_plot}
\alias{dsnb_plot}
\alias{dsnb_plot}
\title{The Stopped Negative Binomial P.M.F. Plot}
\usage{
dsnb_plot(p, s, t, x, offset)

dsnb_plot(p, s, t, x, offset)
}
\arguments{
\item{p}{the probability of a success on each coin flip.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}

\item{offset}{an offset on the domain of the distribution. This is 
used when getting the conditional distribution where the domain does 
not start at 1.}

\item{p}{the probability of a success on each coin flip.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}

\item{offset}{an offset on the domain of the distribution. This is 
used when getting the conditional distribution where the domain does 
not start at 1.}
}
\value{
a plot of the probability mass function.

a plot of the probability mass function.
}
\description{
The plot of the probability mass function for the Stopped Negative Binomial.

The plot of the probability mass function for the Stopped Negative Binomial.
}
\examples{
dsnb_plot(p=0.2, s=7, t=11)
dsnb_plot(p=0.2, s=7, t=11)
}
