% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb.r
\name{ecsnb}
\alias{ecsnb}
\title{Expected Value of the Conditional SNB Distribution}
\usage{
ecsnb(shape, s, t)
}
\arguments{
\item{shape}{the shape parameters of the beta prior.}

\item{s}{number of successes}

\item{t}{number of failures}
}
\description{
Find the expected size of the conditional SNB distribution with specified 
parameters.
}
\examples{
ecsnb(c(0.5, 0.5), s=7, t=11)
}
