% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{esnb}
\alias{esnb}
\title{Expected Value of the SNB Distribution}
\usage{
esnb(p, s, t)
}
\arguments{
\item{p}{success probability}

\item{s}{number of successes}

\item{t}{number of failures}
}
\description{
Find the expected size of an SNB distribution with specified parameters.
}
\examples{
esnb(p=0.2, s=7, t=11)
}
