% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{expected_sample_size_alt_internal.two_stage_curtail_trial}
\alias{expected_sample_size_alt_internal.two_stage_curtail_trial}
\title{Internal Function}
\usage{
\method{expected_sample_size_alt_internal}{two_stage_curtail_trial}(x)
}
\description{
Internal Function
}
