% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{expected_stage1_sample_size}
\alias{expected_stage1_sample_size}
\title{Expected sample size and SD for Stage 1 of the two-stage design}
\usage{
expected_stage1_sample_size(p, n, r)
}
\arguments{
\item{p}{a vector containing the probability of successful outcomes in 
in Stage 1 (p1) and Stage 2 (p2)}

\item{n}{a vector containing maximum sample sizes planned for Stage 1 (n1) and Stage 2 (n2)}

\item{r}{a vector containing the minimum number of Stage 1 successes to continue to 
Stage 2 (r1) and the minimum number of Stage 2 successes to reject the null hypothesis (r2)}
}
\description{
Mean and standard deviation of the minimum number of patients to make a decision to either 
continue to Stage 2 or else terminate early.
}
\examples{
expected_stage1_sample_size(p = c(.8, .2), n = c(5, 31), r = c(3, 11))
}
