% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{expected_total_sample_size}
\alias{expected_total_sample_size}
\title{Expected curtailed sample size for the two-stage design}
\usage{
expected_total_sample_size(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2)}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2)}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2)}
}
\description{
The expected number of patients who are enrolled and followed to their
endpoint before critical endpoints in Stage 1 and Stage 2 are achieved
}
\examples{
expected_total_sample_size(p = c(.8, .2), n=c(6, 30), r=c(4, 11))
expected_total_sample_size(p = c(.8, .2), n = c(18, 18), r = c(12, 11))
}
