% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb.r
\name{flips_to_zplot_df}
\alias{flips_to_zplot_df}
\title{Format a data.frame for plotting by the zplot function}
\usage{
flips_to_zplot_df(flips)
}
\arguments{
\item{flips}{the seqence of ones and zeros denoting responses and 
non-responses respecitively.}
}
\description{
Format a data.frame for plotting by the zplot function
}
