% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r, R/two-stage-design.R
\name{is.wholenumber}
\alias{is.wholenumber}
\alias{is.wholenumber}
\title{Test for integer}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)

is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the number to test.}

\item{tol}{the tolerance for x being a whole number (default 
\code{.Machine$double.eps ^ 0.5})}

\item{x}{the number to test.}

\item{tol}{the tolerance for x being a whole number (default 
\code{.Machine$double.eps ^ 0.5})}
}
\description{
Test for integer

Test for integer
}
