\name{itemscores}
\alias{itemscores}

\encoding{latin1}

\title{Dataset for illustrating curtailment functions}

\description{
  Dataset containing 1000 simulated responses to 20 items on a mental-health 
  questionnaire. Items are scored 0, 1, 2 or 3. The dataset is used in the examples 
  to illustrate the functions. Observations are classified as 'at risk' with a total 
  score after admistration of item 20, equal to or greater than 19 (i.e., a cut-off 
  score of 19 is used).  
}

\examples{
  summary(itemscores)
}