% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{minimax_design}
\alias{minimax_design}
\title{Evaluate the probability that the maximum sample size is needed for the two-stage design}
\usage{
minimax_design(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2)}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2)}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2)}
}
\description{
Evaluate the minimax probability that the maximum sample size n1+n2 is
required to complete the trial
}
\examples{
minimax_design(p=c(0.8, 0.2), n=c(3,33), r=critical_values(n=c(3,33), 
  p=c(0.8, 0.2), pearly=0.1, alpha=0.1))
}
