% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{minimax_probability_internal.two_stage_curtail_trial}
\alias{minimax_probability_internal.two_stage_curtail_trial}
\title{Internal Function}
\usage{
\method{minimax_probability_internal}{two_stage_curtail_trial}(x)
}
\description{
Internal Function
}
