% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{optimal_design}
\alias{optimal_design}
\title{Find the optimal design}
\usage{
optimal_design(x)
}
\arguments{
\item{x}{a two-stage curtailed trial object}
}
\description{
Identifies the optimal design from a trial object containing many 
candidate designs for a given total sample size
}
\examples{
trials <- two_stage_curtail_trial(p1_null = 0.8, p2_null=0.2, 
p1_alt = 0.8, p2_alt = 0.4, n_total=36, prob_early=0.2, alpha=0.05)
optimal_design(trials)
}
