% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2crit}
\alias{ph2crit}
\title{Find critical values (r1, r2)}
\usage{
ph2crit(n, p, pearly = 0.1, alpha = 0.1)
}
\arguments{
\item{n}{n}

\item{p}{p}

\item{pearly}{pearly}

\item{alpha}{alpha}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

