% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2dspb}
\alias{ph2dspb}
\title{Stopped negative binomial distribution mass function}
\usage{
ph2dspb(p, s, t)
}
\arguments{
\item{p}{p}

\item{s}{s}

\item{t}{t}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

