% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2early}
\alias{ph2early}
\title{Probability of stopping early}
\usage{
ph2early(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2) or a scalar containing p1}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2) or a scalar containing n1}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2) or a scalar containing r1}
}
\description{
Calculates the probability of stopping the trial after Stage 1
}
\examples{
ph2early(p = .8, n = 25, r = 18)
ph2early(p = c(.8, .2), n = c(25, 9), r = 18)
}
