% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2Ess}
\alias{ph2Ess}
\title{Expected curtailed sample size}
\usage{
ph2Ess(p, n, r)
}
\arguments{
\item{p}{p}

\item{n}{n}

\item{r}{r}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

