% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2mmax}
\alias{ph2mmax}
\title{Evaluate the probability that the maximum sample size is needed}
\usage{
ph2mmax(p, n, r)
}
\arguments{
\item{p}{p}

\item{n}{n}

\item{r}{r}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

