% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2reject}
\alias{ph2reject}
\title{Probability of rejecting the null hypothesis}
\usage{
ph2reject(p, n, r)
}
\arguments{
\item{p}{p}

\item{n}{n}

\item{r}{r}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

