% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2snb}
\alias{ph2snb}
\title{Stopped negative binomial distribution mass function}
\usage{
ph2snb(p, s, t)
}
\arguments{
\item{p}{probability of success in each trial.}

\item{s}{number of successes.}

\item{t}{number of failures.}
}
\description{
Returns the probability density function for minimum number of 
events for either s successes or t failures.
}
\examples{
# ph2snb(p = .8, s = 3, t = 5)
}
