% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2tnb}
\alias{ph2tnb}
\title{Truncated negative binomial distribution mass function}
\usage{
ph2tnb(p, n, r)
}
\arguments{
\item{p}{p}

\item{n}{n}

\item{r}{r}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

