% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2valid}
\alias{ph2valid}
\title{Check validity of parameter values}
\usage{
ph2valid(p, n, r)
}
\arguments{
\item{p}{p}

\item{n}{n}

\item{r}{r}
}
\description{
DESCRIPTION HERE
}
\examples{
# Put exmple code here.
}

