% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{power_internal.two_stage_curtail_trial}
\alias{power_internal.two_stage_curtail_trial}
\title{Internal function}
\usage{
\method{power_internal}{two_stage_curtail_trial}(x)
}
\description{
Internal function
}
