% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{power_significance_ROC}
\alias{power_significance_ROC}
\title{ROC curve for Power vs. 1-Significance}
\usage{
power_significance_ROC(n, p0, p1, all_labels = FALSE)
}
\arguments{
\item{n}{maximum number of patients in the trial}

\item{p0}{probability of success under the null hypothesis}

\item{p1}{probability of success under the alternative hypothesis}

\item{all_labels}{controls how many labels of s will be included in the plot.  if set to TRUE,  all labels of s will appear 
on the plot. otherwise, if FALSE (default), the labels for the extreme values of s will be removed to improve the readability}
}
\description{
ROC curve of all trial designs with a maximum number of patients, n,
with a varying number of responses, s, required to reach the success endpoint
}
\examples{
power_significance_ROC(17, 0.2, 0.4)
}
