% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{prob_early_stop}
\alias{prob_early_stop}
\title{Probability of stopping early in the two-stage design}
\usage{
prob_early_stop(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2)}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2)}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2)}
}
\description{
Calculates the probability of stopping the trial after Stage 1
}
\examples{
prob_early_stop(p = c(.8, .2), n = c(25, 9), r = c(17, 11))
}
