% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{significance_internal.two_stage_curtail_trial}
\alias{significance_internal.two_stage_curtail_trial}
\title{Internal function}
\usage{
\method{significance_internal}{two_stage_curtail_trial}(x)
}
\description{
Internal function
}
