% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{single_stage_expected_sample_size}
\alias{single_stage_expected_sample_size}
\title{Expected sample size and SD for the one-stage design}
\usage{
single_stage_expected_sample_size(p, s, t)
}
\arguments{
\item{p}{scalar for the one-stage design containing the probability of successful outcome (p)}

\item{s}{number of successes to stop the trial}

\item{t}{number of failures to stop the trial}
}
\description{
Mean and standard deviation of the minimum number of patients to make a decision 
about rejecting the null for the one-stage design.
}
\examples{
single_stage_expected_sample_size(p = 0.2, s = 7, t = 11)
}
