% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{single_stage_power}
\alias{single_stage_power}
\title{Calculate power for the single stage trial}
\usage{
single_stage_power(pAlt, s, t)
}
\arguments{
\item{pAlt}{probability of success under the alternative hypothesis}

\item{s}{number of successes to stop the trial}

\item{t}{number of failures to stop the trial}
}
\description{
Calculates the probability of rejecting the null hypothesis assuming an alternative
probability of success
}
\examples{
single_stage_power(0.5, 7, 11)
}
