% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{stacked_plot}
\alias{stacked_plot}
\title{The Stacked Plot}
\usage{
stacked_plot(x, s, t)
}
\arguments{
\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}
}
\value{
a plot of the probability mass function.
}
\description{
The stacked plot of the probability mass function for the snb showing
the contributions from N (the top barrier) and R (the right barrier).
}
\examples{
stacked_plot(8, 7, 11)
}
