% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{stage1_sample_size_internal.two_stage_curtail_trial}
\alias{stage1_sample_size_internal.two_stage_curtail_trial}
\title{Internal function}
\usage{
\method{stage1_sample_size_internal}{two_stage_curtail_trial}(x)
}
\description{
Internal function
}
