% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{two_stage_critical_values}
\alias{two_stage_critical_values}
\title{For the two-stage design:
Finds r1, the minimum number of Stage 1 successes to continue to 
Stage 2, and r2, the minimum number of Stage 2 successes to 
reject the null hypothesis.  Find r1 so that the probability 
of early stopping is less than or equal to pearly. 
Find r2 from Binomial model with no early stopping
and significance level alpha.}
\usage{
two_stage_critical_values(n, p, pearly = 0.1, alpha = 0.1)
}
\arguments{
\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2).}

\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2) under the null hypothesis.}

\item{pearly}{desired probability of early stopping (default is 0.1).
Not necessary for determining critical values for the one-stage design.}

\item{alpha}{desired significance level (default is 0.1).}
}
\description{
For the two-stage design:
Finds r1, the minimum number of Stage 1 successes to continue to 
Stage 2, and r2, the minimum number of Stage 2 successes to 
reject the null hypothesis.  Find r1 so that the probability 
of early stopping is less than or equal to pearly. 
Find r2 from Binomial model with no early stopping
and significance level alpha.
}
\examples{
two_stage_critical_values(n=c( 5, 31), p=c(.8,.2), pearly=.1, alpha=.1)
}
