% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY-method}
\alias{two_stage_curtail_trial,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY-method}
\title{Create a two_stage_curtail_trial
Case 6:  Anything else}
\usage{

  \S4method{two_stage_curtail_trial}{ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY,ANY}(p1_null,
  p2_null, p1_alt, p2_alt, n1, n2, n_total, r1, r2, prob_early, alpha)
}
\description{
Create a two_stage_curtail_trial
Case 6:  Anything else
}
\examples{
trials <- two_stage_curtail_trial(p_null=c(0.8, 0.2), p_alt=c(0.8, 0.4), 
n_total=36)
}
