% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,missing,numeric,missing,numeric,numeric-method}
\alias{two_stage_curtail_trial,numeric,numeric,missing,numeric,missing,numeric,numeric-method}
\title{Create a two_stage_curtail_trial
Case 4:  User inputs p, n_total, prob_early, alpha}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,missing,numeric,missing,numeric,numeric}(p_null,
  p_alt, n_total, prob_early, alpha)
}
\description{
Create a two_stage_curtail_trial
Case 4:  User inputs p, n_total, prob_early, alpha
}
\examples{
trials <- two_stage_curtail_trial(p_null=c(0.8, 0.2), p_alt=c(0.8, 0.4), 
n_total=36, prob_early=0.1. alpha=0.1)
}
