% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,numeric,missing,numeric,missing,missing-method}
\alias{two_stage_curtail_trial,numeric,numeric,numeric,missing,numeric,missing,missing-method}
\title{Create a two_stage_curtail_trial
Case 1 - User inputs p, n, and r}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,numeric,missing,numeric,missing,missing}(p_null,
  p_alt, n, r)
}
\description{
Create a two_stage_curtail_trial
Case 1 - User inputs p, n, and r
}
\examples{
trial <- two_stage_curtail_trial(p_null=c(0.8, 0.2), p_alt=c(0.8, 0.4), 
n=c(6, 30), r=c(4, 11))
}
