% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,numeric,numeric,missing,missing,numeric,missing,missing,numeric,missing-method}
\alias{two_stage_curtail_trial,numeric,numeric,numeric,numeric,missing,missing,numeric,missing,missing,numeric,missing-method}
\title{Create a two_stage_curtail_trial
Case 4a:  User inputs p, n_total, prob_early}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,numeric,numeric,missing,missing,numeric,missing,missing,numeric,missing}(p1_null,
  p2_null, p1_alt, p2_alt, n_total, prob_early, alpha)
}
\description{
Create a two_stage_curtail_trial
Case 4a:  User inputs p, n_total, prob_early
}
\examples{
trials <- two_stage_curtail_trial(p1_null = 0.8, p2_null=0.2, 
p1_alt = 0.8, p2_alt = 0.4, n_total=36, prob_early=0.1, alpha=0.1)
}
