% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,missing-method}
\alias{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,missing-method}
\title{Create a two_stage_curtail_trial
Case 2:  User inputs p, n
Using default values of alpha and prob_early}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,missing}(p1_null,
  p2_null, p1_alt, p2_alt, n1, n2)
}
\description{
Create a two_stage_curtail_trial
Case 2:  User inputs p, n
Using default values of alpha and prob_early
}
\examples{
trial <- two_stage_curtail_trial(p_null=c(0.8, 0.2), p_alt=c(0.8, 0.4), 
n=c(6, 30))
}
