% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,numeric-method}
\alias{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,numeric-method}
\title{Create a two_stage_curtail_trial
Case 3b:  User inputs p, n, alpha}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,numeric,numeric,numeric,numeric,missing,missing,missing,missing,numeric}(p1_null,
  p2_null, p1_alt, p2_alt, n1, n2, prob_early, alpha)
}
\description{
Create a two_stage_curtail_trial
Case 3b:  User inputs p, n, alpha
}
\examples{
trial <- two_stage_curtail_trial(p_null=c(0.8, 0.2), p_alt=c(0.8, 0.4), 
n=c(6, 30), prob_early=0.1, alpha=0.1)
}
