% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{two_stage_power}
\alias{two_stage_power}
\title{Power of the two-stage design under curtailed sampling}
\usage{
two_stage_power(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2) under the alternative hypothesis}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2)}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2)}
}
\description{
Calculates the probability of rejecting the null hypothesis in the 
two-stage design under curtailed sampling, assuming alternative 
probabililties of success for Stage 1 and Stage 2.
}
\examples{
two_stage_power(p = c(0.8, 0.4), n = c(12, 24), r = c(8, 11))
two_stage_power(p = c(0.8, 0.4), n = c(6, 30), r = c(4, 11))
}
