% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb.r
\name{vcsnb}
\alias{vcsnb}
\title{Variance of the Conditional SNB Distribution}
\usage{
vcsnb(shape, s, t)
}
\arguments{
\item{shape}{the shape parameters of the beta prior.}

\item{s}{number of successes.}

\item{t}{number of failures.}
}
\description{
Find the variance of the conditional SNB distribution with specified 
parameters.
}
\examples{
vcsnb(c(0.5, 0.5), s=7, t=11)
}
