% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curtail.R
\docType{package}
\name{curtail}
\alias{curtail}
\alias{curtail-package}
\title{Design of One- and Two-Stage Clinical Trials with Curtailed Sampling}
\references{
Rucker G (1989). "A two-stage trial design for testing treatment,
self-selection and treatment preference effects." \emph{Stat Med}, 
\strong{8}(4):477-485. 
(\href{https://www.ncbi.nlm.nih.gov/pubmed/2727471}{PubMed})
}
