% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{ph2valid}
\alias{ph2valid}
\title{Check validity of parameter values p, n, and r.}
\usage{
ph2valid(p, n, r)
}
\arguments{
\item{p}{vector containing the probability of successful outcomes
in Stage 1 (p1) and Stage 2 (p2). Must have 0 <= p2 <= p1 <= 1.}

\item{n}{vector containing sample sizes planned for Stage 1 (n1) 
and Stage 2 (n2).  Both n1 and n2 should be non-negative integers.}

\item{r}{vector containing the minimum number of Stage 1 successes 
to continue to Stage 2 (r1) and the minimum number of Stage 2 
successes to reject the null hypothesis (r2). Must have 0 <= r1 <= r2,
0 <= r1 <= n1, and 0 <= r2 <= n1 + n2.}
}
\description{
Check validity of parameter values p, n, and r.
}
\examples{
# ph2valid(p = c(.2, .1), n = c(10, 10), r = c(5, 5))
# ph2valid(p = c( .2, .3), n = c(10, 10), r = c(5, 5))
}
