% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2.r
\name{power_significance_plot}
\alias{power_significance_plot}
\title{Plot power and significance}
\usage{
power_significance_plot(n, p0, p1)
}
\arguments{
\item{n}{maximum number of patients in the trial}

\item{p0}{probability of success under the null hypothesis}

\item{p1}{probability of success under the alternative hypothesis}
}
\description{
Plot power and significance across all trial designs with a maximum number of patients, n,
with a varying number of responses, s, required to reach the success endpoint
}
\examples{
power_significance_plot(17, 0.2, 0.4)
}
