% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-stage-design.r
\name{single_stage_curtail_trial}
\alias{single_stage_curtail_trial}
\title{Generic function for creating single-stage curtail trials}
\usage{
single_stage_curtail_trial(p_null, p_alt, s, t, n)
}
\description{
Generic function for creating single-stage curtail trials
}
