% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\docType{methods}
\name{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,numeric,numeric,missing,missing-method}
\alias{two_stage_curtail_trial,numeric,numeric,numeric,numeric,numeric,numeric,missing,numeric,numeric,missing,missing-method}
\title{Case 1 - User inputs p, n, and r
trial <- two_stage_curtail_trial(p1_null = 0.8, p2_null = 0.2, 
p1_alt = 0.8, p2_alt = 0.4, n1 = 6, n2 = 30, r1 = 4, r2 = 11)}
\usage{

  \S4method{two_stage_curtail_trial}{numeric,numeric,numeric,numeric,numeric,numeric,missing,numeric,numeric,missing,missing}(p1_null,
  p2_null, p1_alt, p2_alt, n1, n2, r1, r2)
}
\description{
Case 1 - User inputs p, n, and r
trial <- two_stage_curtail_trial(p1_null = 0.8, p2_null = 0.2, 
p1_alt = 0.8, p2_alt = 0.4, n1 = 6, n2 = 30, r1 = 4, r2 = 11)
}
