% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{two_stage_curtail_trial}
\alias{two_stage_curtail_trial}
\title{Generic function for creating single-stage curtail trials}
\usage{
two_stage_curtail_trial(p_null, p_alt, n, n_total, r, prob_early, alpha)
}
\description{
Generic function for creating single-stage curtail trials
}
