% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-design.R
\name{two_stage_curtail_trial}
\alias{two_stage_curtail_trial}
\title{Generic function for creating single-stage curtail trials}
\usage{
two_stage_curtail_trial(p1_null, p2_null, p1_alt, p2_alt, n1, n2, n_total, r1,
  r2, prob_early, alpha)
}
\description{
Generic function for creating single-stage curtail trials
}
